Test framework now checks for memory leaks. Renamed some variables to prevent compiler errors on Windows platforms.

+++ duck/arithmetic.c
@@ -240,7 +240,7 @@ VALUE Divide(VALUE a, VALUE b)
 VALUE Modulus(VALUE a, VALUE b)
 {
     VALUE result;
-	int base;
+	int64 base;
     result.type = VAL_NIL;
     result.data.primitive = 0;
